LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/drawing-rectangles-to-the-canvas/2d.strokeRect.globalalpha-expected.txt

2d.strokeRect.globalalpha

strokeRect is affected by globalAlpha


PASS strokeRect is affected by globalAlpha 
